Stranne.BooliLib/Converters/ShortDateJsonConverter.cs
using System;
using Newtonsoft.Json;

namespace Stranne.BooliLib.Converters
{
    internal class ShortDateJsonConverter : JsonConverter
    {
        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
        {
            if (value == null)
            {
                return;
            }

            var date = (DateTime)value;
            writer.WriteValue(date.ToString("yyyyMMdd"));
        }

        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
        {
            throw new NotImplementedException();
        }

        public override bool CanRead => false;

        public override bool CanConvert(Type objectType)
        {
            return objectType == typeof(DateTime) || objectType == typeof(DateTime?);
        }
    }
}
